separated the Claw.java class to anther package and created ClawConstants.java. organized the elevator subsystem and added the ability to change the PID values removed unnecessary files
