+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robotSubSystems/claw/Claw.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.robotSubSystems.elevator;
+package org.firstinspires.ftc.teamcode.robotSubSystems.claw;
 
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.Servo;
@@ -6,7 +6,6 @@ import com.qualcomm.robotcore.hardware.Servo;
 public class Claw {
 
     private Servo clawServo;
-    private boolean lastState = false;
     private boolean isClawOpen;
     public void init(HardwareMap hardwareMap) {
         clawServo = hardwareMap.get(Servo.class, "clawServo");
@@ -14,10 +13,10 @@ public class Claw {
     public void operate(boolean openClaw) {
 
         if (openClaw) {
-            clawServo.setPosition(0.57);
+            clawServo.setPosition(ClawConstants.clawOpenPosition);
             isClawOpen = true;
         } else {
-            clawServo.setPosition(1);   
+            clawServo.setPosition(ClawConstants.clawClosedPosition);
             isClawOpen = false;
         }
     }
