+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robotSubSystems/claw/ClawConstants.java
@@ -1,0 +1,7 @@
+package org.firstinspires.ftc.teamcode.robotSubSystems.claw;
+
+public class ClawConstants {
+
+    public static double clawOpenPosition = 0.56;
+    public static double clawClosedPosition = 1;
+}
