+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robotSubSystems/elevator/ElevatorConstants.java
@@ -2,11 +2,16 @@ package org.firstinspires.ftc.teamcode.robotSubSystems.elevator;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.HardwareMap;
+import com.qualcomm.robotcore.hardware.PIDFCoefficients;
 
 import org.firstinspires.ftc.teamcode.TeleOp14872;
 import org.firstinspires.ftc.teamcode.control.PID;
 
 public class ElevatorConstants {
-    public static PID setElevatorLevelPID = new PID(0,0,0,0,0);
+
     public static double motorRotationsPerLevel = 25.4/11.2;
+    public static int[] elevatorLevelsInTicks = {0, 70, 1815, 2910, 4115};
+    //TODO: find the best values.
+    public static PIDFCoefficients pidfCoefficients = new PIDFCoefficients(0, 0, 0, 0);
+
 }
